Prevent Constant propigation errors and add Mockito.
Prevent a constant propigation error which was occuring because
the consensus was being returned as null. Removed a "debug point"
from Caleb ('wtf?') and added a test case. Needed to import the
dependancy of Mockito to the project so I could Mock out
the surrounding code for a proper unit test, also needed to add
in some dependancy inject to get away from using ConstantBuilder
only as a static class.