Pass around virtual methods rather than strings
This is because profiling the code showed there's a lot of
splitting and other string manipulation functions happening.
Method signatures are passed around, information is looked up
about them, and then thrown away, and later that same info
would be looked up again.

This also cleans up ClassManager, which was getting kind of complex.
And it simplifies lots of other common method related lookups such
as parameter types.