Add VirtualGenerics
This is a major refactor and also includes some unrelated changes.

This change makes it such that smalivm doesn't pass around strings
for classes, methods, and fields constantly. Instead, these are
wrapped up nicely by a VirtualGenerics subtype. This also allowed
a lot of ClassManager functionality to be pushed down to these
objects like VirtualClass and VirtualMethod. Overall, this has the
effect of simplifying a lot of the code and making it easier to
reason about. A minor bonus is that during profiling, a lot of
time is spent doing string operations (e.g. split("->"))