Finish removing VirtualException
The idea behind VirtualException was to act as a wrapper for real exceptions and to make
equality checking easier for testing and in general. Also, this is where stack trace
generation was planned to go. This has a major flaw though, exposed only after
implementing throw and move-exception properly: the VirtualException instances
are going to get put into registers unless you're careful. You'd have to unwrap
the actual exception whenever it was stored in a method state and rewrap it
whenever smalivm wanted to pass it around and look at it. Rather than adding code
everywhere to do this, I just pass around Throwable objects which are created by
the ExceptionFactory which will eventually populate false stack traces to look
legit.

This made some unit tests more onerous to mock, so I removed them in favor of
a more integration test. I believe this is much more clear to read and understand,
though the tests aren't as fast and they're not sweet mocked unit tests.