Improve class generation
This adds support for inner classes, enclosing methods and source attributes when generating classes. It's still not exaclty like Java, but it's a bit better.

Also, this temporarily removes support for class signatures when generating classes. This shouldn't affect 99% of classes and for the classes it does affect, it shouldn't matter most of the time. More research is needed to know exactly how to handle signatures since there are no working examples anywhere in the world on how to do it...